fix the clock type

+++ internal/interpreter/std.go
@@ -6,7 +6,7 @@ import (
 )
 
 func StdFnTime(ctx context.Context, interpeter *interpreter) (any, error) {
-	return time.Now().UnixMilli(), nil
+	return float64(time.Now().UnixMilli()), nil
 }
 
 func StdFnPPrint(ctx context.Context, interpeter *interpreter, args ...any) (any, error) {
